refactor: move fixtures to conftest
