+++ queue_services/entity-digital-credentials/tests/conftest.py
@@ -1,0 +1,121 @@
+# Copyright © 2023 Province of British Columbia
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""Fixutres for the digital credentials queue are contained here."""
+
+import os
+
+import pytest
+from flask import Flask
+from flask_migrate import Migrate, upgrade
+from legal_api import db as _db
+from sqlalchemy import event, text
+from sqlalchemy.schema import MetaData
+
+from entity_digital_credentials.config import get_named_config
+
+
+# Fixtures
+@pytest.fixture(scope='session')
+def app():
+    """Return a session-wide application configured in TEST mode."""
+    _app = Flask(__name__)
+    _app.config.from_object(get_named_config('testing'))
+    _db.init_app(_app)
+
+    return _app
+
+
+@pytest.fixture(scope='session')
+def db(app):  # pylint: disable=redefined-outer-name, invalid-name
+    """Return a session-wide initialised database.
+
+    Drops all existing tables - Meta follows Postgres FKs
+    """
+    with app.app_context():
+        # Clear out any existing tables
+        metadata = MetaData(_db.engine)
+        metadata.reflect()
+        metadata.drop_all()
+        _db.drop_all()
+
+        sequence_sql = """SELECT sequence_name FROM information_schema.sequences
+                          WHERE sequence_schema='public'
+                       """
+
+        sess = _db.session()
+        for seq in [name for (name,) in sess.execute(text(sequence_sql))]:
+            try:
+                sess.execute(text('DROP SEQUENCE public.%s ;' % seq))
+                print('DROP SEQUENCE public.%s ' % seq)
+            except Exception as err:  # pylint: disable=broad-except; # noqa: B902
+                print(f'Error: {err}')
+        sess.commit()
+
+        # ############################################
+        # There are 2 approaches, an empty database, or the same one that the app will use
+        #     create the tables
+        #     _db.create_all()
+        # or
+        # Use Alembic to load all of the DB revisions including supporting lookup data
+        # This is the path we'll use in legal_api!!
+
+        # even though this isn't referenced directly, it sets up the internal configs that upgrade needs
+        legal_api_dir = os.path.abspath('..').replace('queue_services', 'legal-api')
+        legal_api_dir = os.path.join(legal_api_dir, 'migrations')
+        Migrate(app, _db, directory=legal_api_dir)
+        upgrade()
+
+        return _db
+
+
+@pytest.fixture(scope='function')
+def session(app, db):  # pylint: disable=redefined-outer-name, invalid-name
+    """Return a function-scoped session."""
+    with app.app_context():
+        conn = db.engine.connect()
+        txn = conn.begin()
+
+        options = dict(bind=conn, binds={})
+        sess = db.create_scoped_session(options=options)
+
+        # For those who have local databases on bare metal in local time.
+        # Otherwise some of the returns will come back in local time and unit tests will fail.
+        # The current DEV database uses UTC.
+        sess.execute("SET TIME ZONE 'UTC';")
+        sess.commit()
+
+        # establish  a SAVEPOINT just before beginning the test
+        # (http://docs.sqlalchemy.org/en/latest/orm/session_transaction.html#using-savepoint)
+        sess.begin_nested()
+
+        @event.listens_for(sess(), 'after_transaction_end')
+        def restart_savepoint(sess2, trans):  # pylint: disable=unused-variable
+            # Detecting whether this is indeed the nested transaction of the test
+            if trans.nested and not trans._parent.nested:  # pylint: disable=protected-access
+                # Handle where test DOESN'T session.commit(),
+                sess2.expire_all()
+                sess.begin_nested()
+
+        db.session = sess
+
+        sql = text('select 1')
+        sess.execute(sql)
+
+        yield sess
+
+        # Cleanup
+        sess.remove()
+        # This instruction rollsback any commit that were executed in the tests.
+        txn.rollback()
+        conn.close()
